Refactor sperate main.dart and my_app.dart
