+++ test/widget_test.dart
@@ -6,7 +6,7 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter/material.dart';
-import 'package:flutter_tap_color_changer/main.dart';
+import 'package:flutter_tap_color_changer/my_app.dart';
 import 'package:flutter_test/flutter_test.dart';
